- TodaySessions SQL 조건 및 정렬 개선(COALESCE 적용) - started_at 저장 실패 로그 확인 및 SUBSCRIBE 흐름 점검 - SessionEndEvent 역직렬화 문제 해결(NoArgsConstructor 등 적용)

+++ spring-chat-matching/src/main/java/com/chatmatchingservice/springchatmatching/domain/chat/entity/ChatSession.java
@@ -46,7 +46,7 @@ public class ChatSession {
     private LocalDateTime endedAt;
 
     @Column(name = "duration_sec")
-    private Integer durationSec;
+    private Long durationSec;
     // 🔥 추가: DB 스키마에 존재하는 started_at 컬럼
 
     // --- 생성 메서드 ---
